Export platforms in separate file
